Deploy demo site to gh-pages
The module run-sequence is necessary to avoid using the same temporary
folder in different deployments. This is intended to be a temporary
solution until the release of Gulp 4.0 which has support for defining
task dependencies in series or in parallel.

Resolves #12.

Built from 9442a89a5d5fa4b9f3527f509522f634369ddb2e.